refactor: MVC 패턴으로 리팩토링
- MVC 구조 생성
