+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,0 +1,3 @@
+package lotto.controller
+
+class LottoController
