+++ src/main/kotlin/lotto/model/LottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 import camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange
 import lotto.util.ErrorMessage
