+++ src/main/kotlin/lotto/model/LottoResult.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 import lotto.util.LottoConstants
