+++ src/main/kotlin/lotto/model/LottoResultDetail.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 data class LottoResultDetail(
     val lottoRankList: List<LottoRank>,
