+++ src/main/kotlin/lotto/view/LottoInputView.kt
@@ -1,0 +1,3 @@
+package lotto.view
+
+class LottoInputView
