+++ src/main/kotlin/lotto/view/LottoOutputView.kt
@@ -1,0 +1,3 @@
+package lotto.view
+
+class LottoOutputView
