+++ src/test/kotlin/lotto/LottoResultTest.kt
@@ -1,5 +1,8 @@
 package lotto
 
+import lotto.model.Lotto
+import lotto.model.LottoRank
+import lotto.model.LottoResult
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
